backend/requirements.txt
python-dotenv
requests
json
os